requirements.txt
asgiref==3.5.2
black==22.8.0
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
Django==4.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
mypy-extensions==0.4.3
pathspec==0.10.0
platformdirs==2.5.2
pycparser==2.21
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
